backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.3.3
redis==5.0.1
influxdb-client==1.38.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-api-python-client==2.110.0
python-dotenv==1.0.0
cryptography==41.0.7
argon2-cffi==23.1.0
aiofiles==23.2.1
websockets==12.0
jinja2==3.1.2
requests==2.31.0
PyJWT==2.8.0
